+++ src/config-loader.js
@@ -12,8 +12,12 @@ var ConfigLoader = function(options) {
 		}),
 		mediaGenParams: options.mediaGenParams || {}
 	});
+	if (options.verboseErrorMessaging) {
+		this.getErrorMessage = _.identity;
+	}
 	this.initialize.apply(this, arguments);
 },
+
 	template = function(template, data) {
 		template = template.replace(/\{{1,}/g, "{{").replace(/\}{1,}/g, "}}");
 		return _.template(template, data, {
@@ -25,9 +29,10 @@ var ConfigLoader = function(options) {
 		READY: "ready",
 		ERROR: "error"
 	};
+ConfigLoader.DEFAULT_ERROR_MESSAGE = "Sorry, this video is currently not available.";
 ConfigLoader.prototype = {
 	initialize: function() {
-		_.bindAll(this, "onConfigLoaded", "onMediaGenLoaded", "onError");
+		_.bindAll(this, "onConfigLoaded", "onMediaGenLoaded", "onError", "onLoadError");
 		EventEmitter.convert(this);
 	},
 	load: function() {
@@ -36,7 +41,7 @@ ConfigLoader.prototype = {
 		this.request = new Request(
 			url,
 			this.onConfigLoaded,
-			this.onError
+			this.onLoadError
 		);
 	},
 	onConfigLoaded: function(config) {
@@ -45,14 +50,14 @@ ConfigLoader.prototype = {
 			config = config.config;
 		}
 		if (config.error) {
-			this.onError(config.error);
+			this.onError(this.getErrorMessage(config.error));
 			return;
 		}
 		this.config = Config.process(config, this.options);
 		// the config property for the mediaGen can be specified.
 		var mediaGen = this.options.mediaGenURL || config[this.options.mediaGenProperty || "mediaGen"];
 		if (!mediaGen) {
-			this.onError("no media gen specified");
+			this.onError(this.getErrorMessage("no media gen specified."));
 		} else {
 			var mediaGenParams = _.clone(this.options.mediaGenParams);
 			_.each(config.overrideParams, function(value, key) {
@@ -62,7 +67,7 @@ ConfigLoader.prototype = {
 			this.request = new Request(
 				mediaGen,
 				this.onMediaGenLoaded,
-				this.onError
+				this.onLoadError
 			);
 		}
 	},
@@ -72,7 +77,11 @@ ConfigLoader.prototype = {
 			mediaGen = MediaGen.process(mediaGen);
 		} catch (e) {
 			error = true;
-			this.onError(e);
+			if (e.name === MediaGen.MEDIA_GEN_ERROR) {
+				this.onError(e.message);
+			} else {
+				this.onError(this.getErrorMessage(e));
+			}
 		}
 		if (!error) {
 			this.config.mediaGen = mediaGen;
@@ -83,12 +92,18 @@ ConfigLoader.prototype = {
 			});
 		}
 	},
+	onLoadError: function(data) {
+		this.onError(this.getErrorMessage(data));
+	},
 	onError: function(data) {
 		this.emit(Events.ERROR, {
 			type: Events.ERROR,
 			data: data,
 			target: this
 		});
+	},
+	getErrorMessage: function() {
+		return ConfigLoader.DEFAULT_ERROR_MESSAGE;
 	},
 	destroy: function() {
 		if (this.request) {